fix: address comments

+++ scripts/tests.sh
@@ -20,8 +20,11 @@ if [[ -d "tests/" ]]; then
   pip install pytest pytest-mock
 
   if [[ -f "tests/requirements.txt" ]]; then
-  pip install -r tests/requirements.txt
-  pip install .
+    pip install -r tests/requirements.txt
+  fi
+
+  if [[ -f "setup.py" ]]; then
+    pip install .
   fi
   
   pip install -r requirements.txt
